correction de tagRepositoryImpl et de TagModule
j'ai corrige les fichiers tagRepositoryImpl et TagModule

+++ app/src/main/java/com/esmt/projet/victodo/feature_tag/data/data_source/TagDao.kt
@@ -10,7 +10,7 @@ interface TagDao {
     fun getTags(): Flow<List<Tag>>
 
     @Query("SELECT * FROM tag WHERE id= :id")
-    suspend fun getTagById(id: Int): Tag?
+    suspend fun getTagById(id: Long): Tag?
 
     @Insert(onConflict = OnConflictStrategy.REPLACE)
     suspend fun insertTag(tag: Tag)
